+++ src/App.tsx
@@ -2,7 +2,9 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { ToastContainer } from 'react-toastify';
 
-import Home from './pages/Main';
+import Home from './pages/Home';
+import { Users } from './pages/Users';
+import { routes } from './routes';
 
 import './App.css';
 
@@ -11,11 +13,15 @@ const queryClient = new QueryClient();
 const App = () => {
   return (
     <div className="App">
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<Home />} />
-        </Routes>
-      </BrowserRouter>
+      <QueryClientProvider client={queryClient}>
+        <ToastContainer />
+        <BrowserRouter>
+          <Routes>
+            <Route path={routes.home} element={<Home />} />
+            <Route path={routes.users} element={<Users />} />
+          </Routes>
+        </BrowserRouter>
+      </QueryClientProvider>
     </div>
   );
 };
