+++ src/pages/Users.tsx
@@ -1,0 +1,9 @@
+import { Header } from '../components/Header';
+import UserList from '../components/UserList';
+
+export const Users = () => (
+  <>
+    <Header />
+    <UserList />
+  </>
+);
