+++ src/pages/index.ts
@@ -1,2 +1,2 @@
-export * from './Home';
-export * from './Main'
+export * from './Users';
+export * from './Home'
