+++ src/routes.ts
@@ -1,4 +1,5 @@
 export const routes = {
+  home: '/',
   users: '/users',
-  usersPosts: '/usersPosts',
+  usersPosts: '/posts',
 };
